+++ src/app/homepage-module/components/homepage.component.ts
@@ -1,7 +1,7 @@
 import { Component, ViewEncapsulation } from '@angular/core';
 import { Title } from '@angular/platform-browser';
-import { APP_PAGE_TITLE } from "../../app.settings";
-import { LanguageService } from "../../shared/services/language-service/language.service";
+import { APP_PAGE_TITLE } from '../../app.settings';
+import { LanguageService } from '../../shared/services/language-service/language.service';
 import locale from './homepage.locale.json';
 
 @Component({
@@ -12,6 +12,7 @@ import locale from './homepage.locale.json';
       <h2 class="text-4xl font-semibold mb-1">{{ locale[langService.language].PromoText1 }}</h2>
       <h2 class="text-4xl font-semibold">{{ locale[langService.language].PromoText2 }}</h2>
       <div class="mt-6">
+        <!--@router-->
         <button
           type="button"
           class="text-white bg-blue-700 hover:bg-blue-800 focus:ring-4 focus:ring-blue-300 font-medium rounded-lg text-sm px-4 py-2 mr-2 mb-2 dark:bg-blue-600 dark:hover:bg-blue-700 focus:outline-none dark:focus:ring-blue-800"
@@ -22,7 +23,7 @@ import locale from './homepage.locale.json';
       </div>
     </div>
   `,
-  encapsulation: ViewEncapsulation.None
+  encapsulation: ViewEncapsulation.None,
 })
 export class HomepageComponent {
   protected readonly locale = locale;