+++ src/app/not-found-page/not-found-page.component.ts
@@ -1,8 +1,7 @@
 import { Component, ViewEncapsulation } from '@angular/core';
 import { RouterLink } from '@angular/router';
-
 import { LanguageService } from '../shared/services/language-service/language.service';
-import locale from './not-found.locale.json';
+import locale from './not-found-page.locale.json';
 
 @Component({
   standalone: true,
@@ -10,18 +9,19 @@ import locale from './not-found.locale.json';
   template: `
     <div class="py-9 mx-14">
       <h1 class="text-4xl font-semibold mb-5 text-center">
-        {{ locale[langService.language].Title }}
+        {{ locale[lang.language].Title }}
       </h1>
       <p class="text-md block description text-center mb-10">
-        {{ locale[langService.language].Description }}
+        {{ locale[lang.language].Description }}
       </p>
       <div class="block text-center">
+        <!--@router-->
         <a
           routerLink="/"
           class="cursor-pointer rounded text-sm bg-blue-700 text-white py-2 px-4 outline-none"
           rel="noreferrer noopener"
         >
-          {{ locale[langService.language].Button }}
+          {{ locale[lang.language].Button }}
         </a>
       </div>
     </div>
@@ -29,8 +29,8 @@ import locale from './not-found.locale.json';
   encapsulation: ViewEncapsulation.None,
   imports: [RouterLink],
 })
-export class NotFoundComponent {
+export class NotFoundPageComponent {
   protected readonly locale = locale;
 
-  constructor(protected readonly langService: LanguageService) {}
+  constructor(protected readonly lang: LanguageService) {}
 }