+++ src/app/admin-module/components/create-form-control/create-form-control.component.ts
@@ -1,10 +1,17 @@
 import { Component, Inject } from '@angular/core';
-import { FormControl, Validators } from '@angular/forms';
+import { FormControl, NonNullableFormBuilder, Validators } from '@angular/forms';
 import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
 import sharedLocale from '../../../shared/general.locale.json';
 import { LanguageService } from '../../../shared/services/language-service/language.service';
 import { CreateFormControlData } from '../../models/create-form-control-data.type';
 import locale from './create-form-control.locale.json';
+
+type FieldNameForm = {
+  name: string;
+};
+type FieldNameFormGroup = {
+  [P in keyof FieldNameForm]: FormControl<FieldNameForm[P]>;
+};
 
 @Component({
   selector: 'app-create-form-control',
@@ -13,16 +20,18 @@ import locale from './create-form-control.locale.json';
 export class CreateFormControlComponent {
   protected readonly sharedLocale = sharedLocale;
   protected readonly locale = locale;
-  protected fieldName = new FormControl<string>('', {
-    // TODO pattern - only letters (except czech letters)
-    validators: [Validators.required, Validators.minLength(1)],
-    nonNullable: true,
+  protected readonly fieldNameFormGroup = this.formBuilder.group<FieldNameFormGroup>({
+    name: this.formBuilder.control<string>('', {
+      validators: [Validators.required, Validators.minLength(1), Validators.pattern('^[a-zA-Z]+$')],
+    }),
+    // locale: this.formBuilder.control()
   });
 
   constructor(
     @Inject(MAT_DIALOG_DATA) protected readonly data: CreateFormControlData,
     protected readonly languageService: LanguageService,
-    private readonly dialogRef: MatDialogRef<CreateFormControlComponent>
+    private readonly dialogRef: MatDialogRef<CreateFormControlComponent>,
+    private readonly formBuilder: NonNullableFormBuilder
   ) {}
 
   /**
@@ -30,12 +39,12 @@ export class CreateFormControlComponent {
    * @protected
    */
   protected create() {
-    if (this.fieldName.invalid) {
+    if (this.fieldNameFormGroup.invalid) {
       return;
     }
-    const formControlName = this.fieldName.value;
+    const formControlName = this.fieldNameFormGroup.controls.name.value;
     this.data.onSave(formControlName);
-    this.fieldName.reset();
+    this.fieldNameFormGroup.reset();
     this.dialogRef.close();
   }
 }