+++ CHANGELOG.md
@@ -12,6 +12,10 @@ Run command for (without commit):
 
 ### ?
 
+- convert `CreateFormControlComponent` into form group from form control
+
+### 1.14.1
+
 - use `.html` files for component templates
 - rename component name where component is used as page
 - lazy load `NotFoundPageComponent`